+++ fluent-tests/src/test/java/com/azure/mgmttest/RuntimeTests.java
@@ -24,6 +24,8 @@ import com.azure.core.util.Context;
 import com.azure.core.util.serializer.SerializerAdapter;
 import com.azure.core.util.serializer.SerializerEncoding;
 import com.azure.identity.EnvironmentCredentialBuilder;
+import com.azure.json.JsonProviders;
+import com.azure.json.JsonWriter;
 import com.azure.mgmtlitetest.advisor.AdvisorManager;
 import com.azure.mgmtlitetest.advisor.models.ResourceRecommendationBase;
 import com.azure.mgmtlitetest.advisor.models.SuppressionContract;
@@ -34,6 +36,7 @@ import com.azure.mgmtlitetest.mediaservices.MediaServicesManager;
 import com.azure.mgmtlitetest.mediaservices.models.MediaService;
 import com.azure.mgmtlitetest.mediaservices.models.StorageAccountType;
 import com.azure.mgmtlitetest.resources.ResourceManager;
+import com.azure.mgmtlitetest.resources.fluent.models.GenericResourceInner;
 import com.azure.mgmtlitetest.resources.models.ResourceGroup;
 import com.azure.mgmtlitetest.storage.StorageManager;
 import com.azure.mgmtlitetest.storage.models.AccessTier;
@@ -73,6 +76,7 @@ import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 import java.io.File;
 import java.io.IOException;
+import java.io.StringWriter;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
 import java.nio.file.Files;
@@ -498,5 +502,15 @@ public class RuntimeTests {
         Assertions.assertEquals("myName", galleryInner.name());
         Assertions.assertEquals("myDisclaimer", galleryInner.disclaimer());
         Assertions.assertEquals("abc", pirCommunityGalleryResource.uniqueId());
+    }
+
+    @Test
+    public void testWriteUntypedNull() throws IOException {
+        StringWriter stringWriter = new StringWriter();
+        try (JsonWriter writer = JsonProviders.createWriter(stringWriter)) {
+            GenericResourceInner resourceInner = new GenericResourceInner();
+            resourceInner.toJson(writer);
+        }
+        Assertions.assertFalse(stringWriter.toString().contains("properties"));
     }
 }